% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm_to_com.R
\encoding{UTF-8}
\name{arm_to_com}
\alias{arm_to_com}
\title{Convertit les arrondissements municipaux en communes}
\usage{
arm_to_com(
  data,
  from = NULL,
  to = NULL,
  extra = function(x) {
     return(x)
 }
)
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne initiale des
communes ou  arrondissements municipaux.
Par défaut, première colonne. Sans objet si \code{data} est un vecteur.}

\item{to}{Colonne finale pour les communes.
Par défaut, même nom que la colonne initiale. Sans objet si \code{data} est un vecteur.}

\item{extra}{Autres codes géographiques : \code{NULL}, valeur unique, vecteur nommé ou fonction.
Par défaut, les codes géographiques en dehors Paris, Lyon et Marseille ne sont pas changés.}
}
\value{
Un objet du même type que \code{data}.
\itemize{
\item Pour une data frame, une data frame avec le même nombre de lignes.
\item Pour un vecteur, un vecteur de dimension égale.
}
}
\description{
Convertit les codes géographiques des arrondissements municipaux de Paris,
Lyon et Marseille en codes de ces communes.
}
\examples{
x <-  c("01123","13201","13202","75101")

# data frame
data <- data.frame(ID = c(1:4), CODE_ARM = x)
data |> arm_to_com(from = CODE_ARM, to = "CODE_COM")

# vecteur
arm_to_com(x)
arm_to_com(x, extra = NULL)
unique(arm_to_com(x))
}
