% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_cog.R
\encoding{UTF-8}
\name{change_cog}
\alias{change_cog}
\title{Change l'année du code officiel géographique des communes}
\usage{
change_cog(
  data,
  cog_from,
  cog_to,
  from = NULL,
  to = NULL,
  infos = FALSE,
  split_ratio = FALSE,
  one_to_one = FALSE
)
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{cog_from}{Année initiale du code officiel géographique, à partir de 2008.}

\item{cog_to}{Année finale du code officiel géographique, à partir de 2008.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne initiale des communes.
Par défaut, première colonne. Sans objet si \code{data} est un vecteur.}

\item{to}{Colonne finale pour les communes.
Par défaut, même nom que la colonne initiale. Sans objet si \code{data} est un vecteur.}

\item{infos}{Ajouter les informations de la table de passage. Par défaut, FALSE.
Sans objet si \code{data} est un vecteur.}

\item{split_ratio}{Ajouter une clé de répartition pour la réaffectation
des effectifs quand des communes sont scindées. Par défaut, FALSE.
Sans objet si \code{data} est un vecteur.}

\item{one_to_one}{Ne garder qu'une seule commmune en cas de scission.
Le code est celui de la commune initiale s'il est
présent dans l'une des communes issues de la scission, sinon il correspond à
celui de la commune fille qui est la plus peuplée.
Cette option permet d'éviter que des lignes soient dupliquées en cas de scission.
Par défaut, FALSE.}
}
\value{
Un objet du même type que \code{data}.
\itemize{
\item Pour une data frame, une data frame avec un nombre de lignes égal ou supérieur.
Le nombre de ligne est toujours égal si \code{one_to_one = TRUE}.
\item Pour un vecteur, un vecteur de dimension égale ou supérieure.
La dimension est toujours égale si \code{one_to_one = TRUE}.
}
}
\description{
Convertit les codes géographiques des communes dans une autre année du code officiel géographique (COG).
Les années vont de 2008 au dernier millésime disponible.
}
\examples{
# Un exemple de data frame avec quelques communes
data <- data.frame(COM=c("14712", "16233", "16351", "53239", "53249", "53274"))

# Change l'annee du code officiel geographique des communes
data |> change_cog(from = "COM", cog_from = 2019, cog_to = 2021)

# Variante : ne retient qu'une commune apres scission
data |> change_cog(
  from = "COM",
  to = "COM_21",
  cog_from = 2019,
  cog_to = 2021,
  one_to_one = TRUE
)

# Informations de la table de passage
data |> change_cog(
  from = "COM",
  to = "COM_21",
  cog_from = 2019,
  cog_to = 2021,
  infos = TRUE
)

# Cle de repartition pour les scissions
data |> change_cog(
  from = "COM",
  to = "COM_21",
  cog_from = 2019,
  cog_to = 2021,
  split_ratio = TRUE
)

# Pour un vecteur
change_cog(data = c("14712", "16233", "16351"), cog_from = 2019, cog_to = 2021)

}
