% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cog.R
\encoding{UTF-8}
\name{check_cog}
\alias{check_cog}
\title{Vérifie le code géographique des communes}
\usage{
check_cog(
  data,
  cog,
  from = NULL,
  complete = FALSE,
  ignore_arm = TRUE,
  ignore_mayotte = FALSE,
  data_res = FALSE,
  message = TRUE
)
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{cog}{Année du code officiel géographique, à partir de 2008.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne initiale des communes.
Par défaut, première colonne. Sans objet si \code{data} est un vecteur.}

\item{complete}{Vérifier que les données sont complète,
c'est-à-dire que toutes les communes du COG sont représentées.
Par défaut, \code{FALSE} : vérifie uniquement si toutes les communes sont dans le COG.}

\item{ignore_arm}{Ignorer les arrondissements municipaux de Paris, Lyon et Marseille.
Par défaut, TRUE. Si \code{FALSE}, les arrondissements municipaux ne sont pas dans le COG.}

\item{ignore_mayotte}{Ignorer communes de Mayotte.
Par défaut, \code{FALSE.} : les communes de Mayotte sont hors du COG avant 2012 et
comprises dans le COG à partir de 2012.}

\item{data_res}{Renvoyer le resultat de la comparaison dans
une data frame. Par défaut, \code{FALSE.}}

\item{message}{Générer un avertissement quand des différences
avec le COG sont détectées. Par défaut, \code{TRUE}.}
}
\value{
Un booléen (si \code{data_res} vaut \code{FALSE}) ou une data frame (si \code{data_res} vaut \code{TRUE}).
}
\description{
Compare les codes des communes au code officiel geographique (COG) pour une année demandée.
Les années vont de 2008 au dernier millésime disponible.
}
\examples{
data <- data_com(2018)

# Pas d'erreurs
check_cog(data, cog = 2018)

# Messages d'erreur
check_cog(data, cog = 2019)

# Verifier aussi les communes manquantes
check_cog(data, cog = 2019, complete = TRUE)

# Ne pas ignorer les arrondissements municipaux
data <- com_to_arm(data, to = "CODE_ARM")
check_cog(data, cog = 2019, from = CODE_ARM, ignore_arm = FALSE)

# Renvoyer les resultats dans une base
data_errors <- data |>
   check_cog(cog = 2009, complete = TRUE, data_res = TRUE, message = FALSE)

# Pour un vecteur
x <- data$COM
check_cog(x, cog = 2018)

}
