% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_many.R
\encoding{UTF-8}
\name{codes_to_many}
\alias{codes_to_many}
\title{Merge or split geographic codes}
\usage{
codes_to_many(
  data,
  codes_ini,
  codes_fin,
  extra = NULL,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame or vector}

\item{codes_ini}{Initial geographic codes}

\item{codes_fin}{Final geographic codes}

\item{extra}{Other geographic codes}

\item{from}{Initial column (in a data frame)}

\item{to}{Final column (in a data frame)}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame with an equal or greater number of rows
or a vector with an equal or greater dimension.
}
\description{
Convert geographic codes by managing splits and merges.
The number of rows in a data frame or the dimension of a vector are not preserved.
}
\examples{
x <- c("B","C1","C2","A","Z")
codes_1 <- c("A","A","B","C1","C2")
codes_2 <-  c("a1","a2","b","c","c")

# Data frame
dt <- data.frame(X = x)
codes_to_many(dt, from = "X", to = "Y", codes_ini = codes_1, codes_fin = codes_2)

# Vector
codes_to_many(x, codes_ini = codes_1, codes_fin = codes_2)

}
