% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_events.R
\encoding{UTF-8}
\name{cog_events}
\alias{cog_events}
\title{Évènements du code officiel géographique}
\usage{
cog_events(x, message = TRUE)
}
\arguments{
\item{x}{Code géographique.}

\item{message}{Générer un avertissement si aucun évènement n'a eu lieu
depuis 2008 ou si le code demandé n'aparaît pas dans le code officel
géographique. Par défaut, TRUE.}
}
\value{
Une data frame
}
\description{
Renvoie une data frame avec les évenements ayant eu lieu depuis 2008 pour un
code géographique donné : fusions, scissions ou changement de code.
}
\details{
Colonnes de la data frame générée par la fonction \code{cog_events()} :
\itemize{
\item \code{COG_INI} : Année initiale du code officiel géographique.
\item \code{COG_FIN} : Année finale du code officiel géographique.
\item \code{COM_INI} : Code initial de la commune.
\item \code{COM_FIN} : Code final de la commune.
\item \code{NB_COM_INI} : Nombre initial de communes, \emph{supérieur à 1 pour une fusion}.
\item \code{NB_COM_FIN} : Nombre final de communes, \emph{supérieur à 1 pour une scission}.
}
}
\examples{
# Exemple d'une commune avec un changement de code et une fusion
cog_events("14472")

# Exemple d'une commune avec une fusion et un retablissement (scission)
cog_events("14712")

# Exemple d'une commune sans evenements dans le COG
cog_events("13001")

# Exemple d'un code n'aparaissant pas dans le COG
cog_events("13999")

}
