% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{cog_ref}
\alias{cog_ref}
\title{Année de la base communale de référence}
\format{
Nombre
}
\usage{
cog_ref
}
\description{
Année de la base communale de référence
}
\keyword{datasets}
