% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_transition.R
\encoding{UTF-8}
\name{cog_transition}
\alias{cog_transition}
\title{Table de passage entre deux années du code officiel géographique des communes}
\usage{
cog_transition(cog_from, cog_to)
}
\arguments{
\item{cog_from}{Année du code officiel géographique des communes initiales dans la table de passage}

\item{cog_to}{Année du code officiel géographique des communes finales dans la table de passage}
}
\value{
Une data frame
}
\description{
Renvoie la table de passage des communes qui sont modifiées entre deux années du code officiel géographique (COG).
}
\details{
Colonnes de la data frame générée par la fonction \code{cog_transition()} :
\itemize{
\item \code{COM_INI} : Code commune initial
\item \code{COM_FIN} : Code commune final
\item \code{POP_INI} : Population initiale, pouvant servir de pondération pour la fonction \code{adapt_to_change()}
\item \code{POP_FIN} : Population finale, pouvant servir de pondération pour la fonction \code{adapt_to_change()}
\item \code{NB_COM_INI} : Nombre de communes initial
\item \code{NB_COM_FIN} : Nombre de communes final
}
}
\examples{
cog_transition(cog_from = 2019, cog_to = 2020)

}
