% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_to_arm.R
\encoding{UTF-8}
\name{com_to_arm}
\alias{com_to_arm}
\title{Convertit les communes en arrondissements municipaux}
\usage{
com_to_arm(
  data,
  from = NULL,
  to = NULL,
  extra = function(x) {
     return(x)
 }
)
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne initiale des communes.
Par défaut, première colonne. Sans objet si \code{data} est un vecteur.}

\item{to}{Colonne finale pour les communes ou arrondissements municipaux.
Par défaut, même nom que la colonne initiale. Sans objet si \code{data} est un vecteur.}

\item{extra}{Autres codes géographiques : valeur unique, paires de clés et de valeurs ou fonction.
Par défaut, les codes géographiques en dehors Paris, Lyon et Marseille ne sont pas changés.}
}
\value{
Un objet du même type que \code{data}.
\itemize{
\item Pour une data frame, une data frame avec un nombre de lignes égal ou supérieur.
\item Pour un vecteur, un vecteur de dimension égale ou supérieure.
}
}
\description{
Convertit les codes géographiques des communes de Paris, Lyon et Marseille
en codes géographiques d'arrondissements municipaux.
}
\examples{
x <- c("01123","13055","75056")

# data frame
data <- data.frame(ID = c(1:3), CODE_COM = x)
data |> com_to_arm(from = CODE_COM, to = "CODE_ARM") |> head()

data |> com_to_arm(from = CODE_COM) |> head()

# vecteur
com_to_arm(x)
com_to_arm(x, extra = "?")
}
