% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_to_dep.R
\encoding{UTF-8}
\name{com_to_dep}
\alias{com_to_dep}
\title{Convertit les communes en départements}
\usage{
com_to_dep(
  data,
  from = "COM",
  to = "DEP",
  extra = c(`977` = "977", `978` = "978", `986` = "986", `987` = "987", `988` = "988", ZZ
    = "ZZZ", `NA` = "999")
)
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne initiale des communes.
Par défaut, "COM". Sans objet si \code{data} est un vecteur.}

\item{to}{Colonne finale pour les départements.
Par défaut, "DEP". Sans objet si \code{data} est un vecteur.}

\item{extra}{Autres codes géographiques : \code{NULL}, valeur unique, vecteur nommé ou fonction.
Par défaut, collectivités d'outre-mer et étranger.}
}
\value{
Un objet du même type que \code{data}.
\itemize{
\item Pour une data frame, une data frame avec le même nombre de lignes.
\item Pour un vecteur, un vecteur de dimension égale.
}
}
\description{
Convertit les codes géographiques des communes en codes géographiques des départements.
}
\examples{
x <-  c("84001", "75001", "75001", "97401", "98601", "YYYYY", "99999", "A1001", NA)

# data frame
data <- data.frame(ID = c(1:length(x)), COM = x)
data |> com_to_dep(from = COM, to = "DEP")

# Personalisation des codes extras
codes_extra <- c("977" = "ZZZ", "978" = "ZZZ", "986" = "ZZZ", "987" = "ZZZ",
 "988" = "ZZZ", "ZZ" = "ZZZ", "YY" = "YYY", "NA" = "999")

data |> com_to_dep(from = COM, to = "DEP", extra = codes_extra)

# Vecteur
com_to_dep(x)
com_to_dep(x, extra = NULL)
com_to_dep(x, extra = function(x) {return (x)})

}
