% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{data_arm_com}
\alias{data_arm_com}
\title{Table de passage entre les arrondissements municipaux et les communes de Paris, Lyon et Marseille}
\format{
Une data frame avec 2 variables :
\describe{
\item{COM}{Code commune}
\item{ARM}{Code arrondissement}
}

Pour plus de détails sur le code officiel géographique \url{https://www.insee.fr/fr/information/2560452}
}
\usage{
data_arm_com
}
\description{
Table de passage entre les arrondissements municipaux et les communes de Paris, Lyon et Marseille
}
\keyword{datasets}
