% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_com.R
\encoding{UTF-8}
\name{data_com}
\alias{data_com}
\title{Base des codes des communes selon l'année du code officiel géographique}
\usage{
data_com(cog)
}
\arguments{
\item{cog}{Une année du code officiel géographique des communes (COG).}
}
\value{
Une data frame.
}
\description{
Renvoie une data frame avec les codes géographiques des communes françaises
selon l'année demandée du code officiel géographique (COG).
}
\examples{
data_com_2018 <- data_com(cog = 2018)

}
