% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{data_com_ref}
\alias{data_com_ref}
\title{Base communale de référence}
\format{
Une data frame avec 34945 lignes et 1 variable :
\describe{
\item{COM}{Code géographique en 2023}
}

Pour plus de détails sur le code officiel géographique \url{https://www.insee.fr/fr/information/2560452}
}
\usage{
data_com_ref
}
\description{
Base intermédiaire utilisée par les bases communales antérieures.
}
\keyword{datasets}
