% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{data_dep_reg}
\alias{data_dep_reg}
\title{Table de passage entre les codes des départements et les codes des régions}
\format{
Une data frame avec 2 variables :
\describe{
\item{REG}{Code région}
\item{DEP}{Code département}
}

Pour plus de détails sur le code officiel géographique \url{https://www.insee.fr/fr/information/2560452}
}
\usage{
data_dep_reg
}
\description{
Table de passage entre les codes des départements et les codes des régions
}
\keyword{datasets}
