% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{data_table_passage}
\alias{data_table_passage}
\title{Table de passage du code officiel géographique des communes}
\format{
Une data frame avec 2724 lignes et 34 variables :
\describe{
\item{COM_2008}{Code géographique en 2008}
\item{COM_2009}{Code géographique en 2009}
\item{COM_2010}{Code géographique en 2010}
\item{COM_2011}{Code géographique en 2011}
\item{COM_2012}{Code géographique en 2012}
\item{COM_2013}{Code géographique en 2013}
\item{COM_2014}{Code géographique en 2014}
\item{COM_2015}{Code géographique en 2015}
\item{COM_2016}{Code géographique en 2016}
\item{COM_2017}{Code géographique en 2017}
\item{COM_2018}{Code géographique en 2018}
\item{COM_2019}{Code géographique en 2019}
\item{COM_2020}{Code géographique en 2020}
\item{COM_2021}{Code géographique en 2021}
\item{COM_2022}{Code géographique en 2022}
\item{COM_2023}{Code géographique en 2023}
\item{COM_2024}{Code géographique en 2024}
\item{POP_GEO_2008}{Population de référence en géographie 2008}
\item{POP_GEO_2009}{Population de référence en géographie 2009}
\item{POP_GEO_2010}{Population de référence en géographie 2010}
\item{POP_GEO_2011}{Population de référence en géographie 2011}
\item{POP_GEO_2012}{Population de référence en géographie 2012}
\item{POP_GEO_2013}{Population de référence en géographie 2013}
\item{POP_GEO_2014}{Population de référence en géographie 2014}
\item{POP_GEO_2015}{Population de référence en géographie 2015}
\item{POP_GEO_2016}{Population de référence en géographie 2016}
\item{POP_GEO_2017}{Population de référence en géographie 2017}
\item{POP_GEO_2018}{Population de référence en géographie 2018}
\item{POP_GEO_2019}{Population de référence en géographie 2019}
\item{POP_GEO_2020}{Population de référence en géographie 2020}
\item{POP_GEO_2021}{Population de référence en géographie 2021}
\item{POP_GEO_2022}{Population de référence en géographie 2022}
\item{POP_GEO_2023}{Population de référence en géographie 2023}
\item{POP_GEO_2024}{Population de référence en géographie 2024}
}
}
\usage{
data_table_passage
}
\description{
Base intermédiaire utilisée pour le changement de géographie des communes.
Contient les communes qui ont changé de code officiel géographique depuis 2008.
La population des communes sert de pondération par défaut pour recalculer des variables.
Voir le découpage communal (\url{https://www.insee.fr/fr/information/2028028})
et l'historique des populations légales (\url{https://www.insee.fr/fr/statistiques/2522602})
sur Insee.fr.
}
\keyword{datasets}
