% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_to_reg.R
\encoding{UTF-8}
\name{dep_to_reg}
\alias{dep_to_reg}
\title{Convertit les départements en régions}
\usage{
dep_to_reg(data, from = "DEP", to = "REG", extra = c(`999` = "99", ZZZ = "ZZ"))
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne initiale des départements.
Par défaut, "DEP". Sans objet si \code{data} est un vecteur.}

\item{to}{Colonne finale pour les régions.
Par défaut, "REG". Sans objet si \code{data} est un vecteur.}

\item{extra}{Autres codes géographiques : \code{NULL}, valeur unique, vecteur nommé ou fonction.
Par défaut, collectivités d'outre-mer et étranger.}
}
\value{
Un objet du même type que \code{data}.
\itemize{
\item Pour une data frame, une data frame avec le même nombre de lignes.
\item Pour un vecteur, un vecteur de dimension égale.
}
}
\description{
Convertit les codes géographiques des départements en codes géographiques des régions.
}
\examples{
x <-  c("13", "84", "75", "75", "999", "ZZZ","YYY", NA)

# data frame
data <- data.frame(ID = c(1:length(x)), CODE_DEP = x)
data |> dep_to_reg(from = CODE_DEP, to = "CODE_REG")

# vecteur
dep_to_reg(x)
dep_to_reg(x, extra = c("YYY"="YY"))
unique(dep_to_reg(x))

}
