% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mayotte.R
\encoding{UTF-8}
\name{filter_mayotte}
\alias{filter_mayotte}
\title{Enlève (ou pas) les communes de Mayotte, selon l'année du code officiel géographique}
\usage{
filter_mayotte(data, cog = 2008, from = NULL)
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{cog}{Année du cog. Par défaut, 2008 : les communes de Mayotte sont supprimées.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne des communes.
Par défaut, première colonne. Sans objet si \code{data} est un vecteur.}
}
\value{
Un objet du même type que \code{data}.
\itemize{
\item Pour une data frame, une data frame avec un nombre de lignes inférieur ou égal.
\item Pour un vecteur, un vecteur de dimension inférieure ou égale.
}
}
\description{
Si \code{cog >= 2012}, les communes de Mayotte sont conservées.
Si \code{cog < 2012}, les communes de Mayotte sont supprimées.
}
\examples{
data <- data.frame(COM = c("97424", "97601"))

# Par défaut les lignes des communes de Mayotte sont supprimées
data |> filter_mayotte(from = COM)

# Si cog >= 2012 elles sont conservées
data |> filter_mayotte(cog = 2013, from = COM)

# Pour un vecteur
filter_mayotte(c("97424", "97601"), from = COM)

}
