% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_to_dep.R
\encoding{UTF-8}
\name{reg_to_dep}
\alias{reg_to_dep}
\title{Convertit les régions en départements}
\usage{
reg_to_dep(data, from = "REG", to = "DEP", extra = c(`99` = "999", ZZ = "ZZZ"))
}
\arguments{
\item{data}{Un objet de type data frame ou vecteur.}

\item{from}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} Colonne initiale des régions.
Par défaut, "REG". Sans objet si \code{data} est un vecteur.}

\item{to}{Colonne finale pour les départements.
Par défaut, "DEP". Sans objet si \code{data} est un vecteur.}

\item{extra}{Autres codes géographiques : \code{NULL}, valeur unique, vecteur nommé ou fonction.
Par défaut, collectivités d'outre-mer et étranger.}
}
\value{
Un objet du même type que \code{data}.
\itemize{
\item Pour une data frame, une data frame avec un nombre de lignes égal ou supérieur.
\item Pour un vecteur, un vecteur de dimension égale ou supérieure.
}
}
\description{
Convertit les codes géographiques des régions en codes géographiques des départements.
}
\examples{
x <- c("94", "93", "ZZ", NA)

# data frame
data <- data.frame(ID = c(1:length(x)), CODE_REG = x)
reg_to_dep(data, from = CODE_REG,  to = "CODE_DEP")
# vecteur
reg_to_dep(x)
reg_to_dep(x, extra = "?")
}
