% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_many.R
\encoding{UTF-8}
\name{codes_to_many}
\alias{codes_to_many}
\title{Merge or split geographic codes}
\usage{
codes_to_many(
  data,
  codes_ini,
  codes_fin,
  extra = NULL,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame or vector}

\item{codes_ini}{Initial geographic codes}

\item{codes_fin}{Final geographic codes}

\item{extra}{Other geographic codes}

\item{from}{Initial column (in a data frame)}

\item{to}{Final column (in a data frame)}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame with an equal or greater number of rows
or a vector with an equal or greater dimension.
}
\description{
Convert geographic codes by managing splits and merges.
The number of rows in a data frame or the dimension of a vector are not preserved.
}
\examples{
# Region codes changing in 2016
new_reg <- c("27", "27", "28", "28", "32", "32", "44", "44", "44", "75", "75",
             "75", "76", "76", "84", "84")
old_reg <- c("26", "43", "25", "23", "22", "31", "21", "41", "42", "54", "74",
             "72", "73", "91", "82", "83")

# A data frame with some new regions
data <- data.frame(REG = c("11", "27", "84"))

# Convert into old regions
data |> codes_to_many(
   codes_ini = new_reg,
   codes_fin = old_reg,
   from = "REG",
   to = "OLD_REG",
   extra = function(x){x}
)

# With a vector
codes_to_many(
  data = c("11", "27", "84"),
  codes_ini = new_reg,
  codes_fin = old_reg,
  extra = function(x){x}
)

}
