% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_one.R
\encoding{UTF-8}
\name{codes_to_one}
\alias{codes_to_one}
\title{Merge geographic codes}
\usage{
codes_to_one(
  data,
  codes_ini,
  codes_fin,
  extra = NULL,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame or vector}

\item{codes_ini}{Initial geographic codes}

\item{codes_fin}{Final geographic codes}

\item{extra}{Other geographic codes}

\item{from}{Initial column (in a data frame)}

\item{to}{Final column (in a data frame)}

\item{...}{Arguments passed to or from other methods.}
}
\value{
a data frame with the same number of rows or a vector of equal dimension
}
\description{
Convert geographic codes while preserving the number of rows in a data frame
or the dimension of a vector. Merging is handled, but not splitting.
}
\examples{
# Region codes changing in 2016
new_reg <- c("27", "27", "28", "28", "32", "32", "44", "44", "44", "75", "75",
             "75", "76", "76", "84", "84")
old_reg <- c("26", "43", "25", "23", "22", "31", "21", "41", "42", "54", "74",
             "72", "73", "91", "82", "83")

# A data frame with some old regions
data <- data.frame(REG = c("11", "26", "43", "82", "83"))

# Convert into new regions
data |> codes_to_one(
   codes_ini = old_reg,
   codes_fin = new_reg,
   from = "REG",
   to = "NEW_REG",
   extra = function(x){x}
)

# With a vector
codes_to_one(
  data = c("11", "26", "43", "82", "83"),
  codes_ini = old_reg,
  codes_fin = new_reg,
  extra = function(x){x}
)

}
